Start handling form submission
Check that the domain provided is a Wikimedia domain (so we don’t send
API requests to arbitrary domains), for now using the same regex as in
the frontend (the flavors are sufficiently compatible). Tests are also
added so that we can later improve the implementation with confidence
(in the frontend we’re restricted to regex, in the backend we’re not).

The list of domains is taken from the Subject Alt Name extension of the
certificate used on Wikmedia website.