Add siteinfo module
Since this tool is supposed to work on all Wikimedia wikis, we need to
know what the Category namespace is called (adding [[Category:Foo]]
would always work, but is not nice on non-English wikis) and by what
other names category links may be added (in which case we should not add
them again).

I don’t really like the module name, but it’ll do for now.