Implement AddCategoryAction.apply()
This actually implements the action: it accepts a piece of wikitext
(usaully the full code of a page) and returns it with the category
added, using the MediaWiki Parser from Hell. Some of this code may later
move upwards into CategoryAction so that RemoveCategoryAction can use it
as well.

Some of the tests already verify behavior with sort keys, but this
action doesn’t directly do anything with them; I plan to add that using
separate action subclasses later.