Implement RemoveCategoryAction.apply()
The counterpiece to AddCategoryAction.apply(). Code to check whether a
wikilink is a category link is extracted to CategoryAction, but the rest
works differently: since we need to adjust the text of surrounding nodes
(instead of inserting a new '\n' node before the new wikilink), we
operate directly on the nodes array of the wikicode object.

Just like AddCategoryAction.apply(), this doesn’t yet do anything
special with sort keys (and this is checked by one test), but I plan to
do that using a subclass later.