Add functions to access interface messages to siteinfo
To aggregate individual actions’ summaries into a summary for a full
command, we need the comma-separator and parentheses messages. Currently
we just query the MediaWiki API with each call, but we can add caching
for that (and also category_info()) later. (Of course, for parentheses
we’ll have to cache the version of the message without a message
argumennt, and then replace $1 with the argument ourselves. Fortunately
no translation does anything more complicated with the argument than
simple replacement, and I don’t think that’s going to change either.)

The test messages are taken from zh.wikipedia.org.