Implement Command.run()
This applies all the actions in the command to the page’s wikitext and
then submits the result as a single edit, with the comma-separated edit
summaries of the individual actions (parenthesized if an action was a
no-op) making up the overall edit summary.

No tests yet, because I’m not sure how to test parts of this. The
current FakeSession class won’t cut it, at least.