Extract runner module
This extracts most of the logic from Command.run() into a separate
Runner class and renames the rest to Commany.apply(), a short function
that doesn’t directly deal with the API anymore.

Runner.run_command() is a class method, not just a toplevel function,
because I think the Runner class will be a good place to cache siteinfo
results (we shouldn’t have to ask for it every time).