Rename Batch to NewBatch and remove authentication
The old class wasn’t fully thought through: to properly handle batches,
we need not just authentication data, but also a batch ID and
information which commands have already been run, for instance, because
we don’t typically run a whole batch in one go. Let’s defer all this
complexity to a separate task and instead make this class (the output of
the parser) a very simple container.