Add OpenBatch class
A persistently stored collection of commands, with its own ID, tracking
the status of each command. This is what we’ll store in the database
later on.

The class is certainly not yet complete – we need to store at least the
user who submitted a batch, and if we want to support background
processing, their OAuth credentials as well – but I’ve been
procrastinating on the perfect design of this in my head for almost a
week now, and need to start putting some code out and proceeding based
on my experience with that, rather than empty mental gymnastics. As long
as the tool isn’t deployed yet, it’s not a problem to proceed with this
incomplete version for now.

The command tests are slightly adjusted so that we have more different
commands, otherwise the test for OpenBatch.__str__() isn’t very
effective (both the command plan and finish would have the same str).