Store a single list of command records in NewBatch
Storing two lists is really cumbersome; when we need some commands to
run, we can just scan the one list for CommandPlans: that’s not very
efficient but okay for the in-memory store, which is only intended for
development use anyways.

This also changes the stringification of CommandFinish and subclasses to
put a # in front of the command’s string, which was previously done only
in NewBatch’s stringification.