Add basic HTML batch display
Show a page of batches (with the offset and limit configurable as URL
parameters; links to adjacent pages will be added later), with the
individual commands being defined by templates.

The templates should include links – to the page, and for edits/noops
also to the diff/revision – but I can’t do that because nowhere in the
batch is the wiki domain stored! Oops! That’s a pretty glaring omission
I didn’t notice in all my in-head design :D