Add domain to OpenBatch
When we store a batch, we have to store the domain to act against along
with it. However, there’s no need to have this information in the
NewBatch too (at least not yet), so let’s leave it out of there for now
(otherwise parse_tpsv() needs a domain argument, which feels odd).

Arguably, the domain belongs conceptually with each command or even
action as well, but that feels very wasteful. For rendering them I’ll
just pass in the domain separately.