Add hyperlinks to command templates
Turn the page titles and badges into hyperlinks according to their
attributes, now that we have the domain.

I feel like I want to rip this rendering code apart a bit in the future,
so that there are template functions to render command records, commands
and actions and the surrounding templates calling into them, instead of
app.py’s batch() doing it up-front, but that can wait. (I’ll probably do
it once I try to turn the categories mentioned in the actions into
links, because in the current form that looks a bit painful.)