Turn command rendering into template globals
Instead of pre-rendering the commands and command records and feeding
them into the command record and batch templates respectively, expose
those parts as template globals and have them be called from the
templates. This seems cleaner and more flexible to me.

I’d prefer to make them template filters, not globals (i. e. used as
{{ command_edit.command | render_command }}), but that’s not currently
possible because the domain is not part of the command (record) and has
to be passed as a separate argument. Perhaps I’ll improve that later.