Add assert=user to editing API requests
Toolforge IPs should be blocked from anonymous editing anyways, but just
to make sure, let’s ask the MediaWiki API to throw an error if our
session is somehow not logged in.

Unfortunately, because assert is a keyword in Python, this means we have
to switch from named arguments to spreading a dict of arguments.