Store user identity with batch
An OpenBatch now contains the identity of the user who submitted the
batch, in three formats:

* the user ID on the wiki where the batch will be run
* the globalauth user ID
* the user name, even though it’s not stable, so that we don’t have to
  look it up from the above information whenever we show a batch

The batch store is responsible for extracting this information (along
with the domain) from the mwapi.Session, which it now receives (instead
of just the domain). We also show it when displaying a batch, using
Special:Redirect/user/<id> so that the link is correct even if the user
is renamed (though the name we display will be outdated in that case).

I’m not sure if we’ll use the globalauth user ID, but let’s keep it for
now, it’s not expensive to retrieve.