Implement method to run a slice of a batch
Currently, this manages 50 commands in some 58 seconds. I expect to
improve this a lot, due to the following factors:

* We can cache interface messages, instead of getting them each time we
  want to assemble an edit summary. (We can also get multiple messages
  with one API call.)
* We can cache the CSRF token, instead of getting it for each edit.
* When this tool is deployed on Toolforge, there will be much less
  network delay for API calls.

The result of a command is stored by directly manipulating the
OpenBatch’s list of command records, which is not very nice – in
particular, when a batch is done this results in an OpenBatch of only
finished commands, which is not supposed to exist (there’s an assertion
in the constructor to prevent it, but we’re not using the constructor at
this point). I’ll refactor this once the time comes to introduce a more
persistent, likely SQL-based batch store.