Add pagination to batch page
For now, it’s limited to just “previous” and “next” links, since
individual page numbers don’t really match the current model of a
configurable offset and limit (with offset=75 and limit=50, which page
number is the current one?). Perhaps I’ll change that to a single page
number parameter later, not sure. For the time being, this works.

(I’m slightly worried that it looks a bit like “previous/next batch”
instead of “… page of commands”, but that can also be improved later.)