Cache siteinfo for one day
None of the information we get here should change very often, so we can
comfortably cache it for at least a day (usually much longer, but on the
other hand the API call isn’t *that* expensive). This reduces the time
to run 50 commands from an earlier 58 seconds to some 48 seconds, which
is pretty good already.