Store CSRF token in Runner
Make the Runner bound to one session (changing it from a run_command()
parameter into a constructor parameter), and get the CSRF token once in
the constructor instead of when running each individual command, since
the token is valid for the entire login session. This further reduces
the time to run 50 commands to some 37 seconds.