Get revisions for multiple pages at once
Split the runner into two parts, one to get the information for a page
(wikitext and various metadata) and one to actually edit it. The first
part can be done in bulk, so that method accepts a list of titles and
app.py calls it with all the titles to be processed, which saves a lot
of API calls. This further reduces the time to run 50 commands to some
25 seconds, over twice as fast as at the beginning of the day :)