Start adding DatabaseStore
This implementation stores batches in a MySQL/MariaDB database. At the
moment, loading isn’t supported yet, that will come later.

The database connection is opened each time you store a batch, instead
of being a permanent member of the store, because persistent connections
are not allowed on Toolforge [1].

An integration test is included: if you have a database running on
localhost (default port) and provide the root password in the
MARIADB_ROOT_PASSWORD environment variable, the test harness will create
a temporary database and user and try to store a batch there, then clean
them up afterwards (even if the test failed). We can also test this on
Travis CI (MariaDB 10.1 is the version running on Toolforge).

[1]: https://wikitech.wikimedia.org/wiki/Help:Toolforge/Database#Connection_handling_policy