Remove assertions from OpenBatch constructor
The idea was to have two classes: OpenBatch, with at least one
CommandPlan in it, and ClosedBatch, with only CommandFinishes in it.
That distinction has not yet been fully implemented, but we do now have
situations where the OpenBatch constructor may be called with a list of
only CommandFinishes (when loading it from the database, after all
commands were run – compare the commit message of 996cbdf40b), so for
now it’s better to just disable the assertions.

Also, I think this was potentially loading a whole bunch of commands
from the database, each on a new connection (😱), so I don’t plan to
reintroduce those assertions later even if a ClosedBatch class emerges
after all. Caller discipline will have to suffice there.