Correctly handle case-sensitive categories
Namespaces, including the Category namespace, can be fully case
sensitive (e. g. Wiktionary) or ignore the case of the first letter
(default). Correctly account for this by including the case handling of
the Category namespace in the category info, and applying it to both
category names when checking whether two category links describe the
same category.