Don’t try to run more than 50 commands at once
We let users view up to 500 commands per page, but for editing we don’t
support more than 50 per request (this is currently asserted in
Runner.prepare_pages(), as a consequence of this being the limit on the
MediaWiki API’s titles parameter, but I’m happy with making this an
intentional limit of the tool as well). Instead of crashing if the user
chose “run these commands” on a larger page, stop collecting command
plans after we have 50. The offset and limit parameters are unchanged,
so the user will still see the same page, it just won’t have run all the
commands that are listed.