Detect errors for missing pages
A new general type of CommandFinish is added, CommandFailure, with a
first subclass, CommandPageMissing. The Runner returns this type if a
page is reported as missing in prepare_pages(), the DatabaseStore learns
to serialize and deserialize it, and the app learns to render it.

Just in case it’s important, we record the current timestamp (as
reported by MediaWiki) when the page is missing, though we don’t do
anything with that information at the moment.

The DatabaseStore part isn’t tested yet, because I want to refactor the
serialization/deserialization so that it can be tested without a
database connection. (Database tests are super expensive if we tear down
and destroy whole databases each time!)