Fix handling of non-normalized page titles
If we query the API for non-normalized page titles (e. g. initial
lowercase in “first-letter” namespaces, or using underscores instead of
spaaces), the API will use the normalized title in its response, and
additionally include a map from unnormalized to normalized page names.
If we don’t look at that map, we’ll get a KeyError when looking up the
information by its unnormalized title.

(We could also try to normalize page titles in the tool, but tbh I’d
rather not. MediaWiki knows how to do it correctly, let’s leave it to do
its job instead of implementing our own normalization that probably
wouldn’t be *quite* correct.)