Clean up user-submitted batches before storing them
The “feature” of preserving the combination of spaces and underscores in
the user’s input has turned out to be a giant footgun. Some users *will*
generate input from URLs, and some users *will* not think to replace
underscores with spaces before sending that input to the tool. Since no
one knows of a category that actually *should* be added with
underscores, it’s really not a good look to tell users that they
should’ve looked at the “gotchas”, which were totally documented – let’s
just clean the batch up for them and replace all underscores with
spaces, and be done with it.

However, we only do this on new batch submissions. For better or worse,
there are now some batches on the Toolforge installation that edited
using underscores, and we shouldn’t misrepresent those by cleaning them
up when loading them from the database.