Immediately retry after edit conflicts
If a command failure is safe to retry immediately (currently, this is
only the case for edit conflicts), do so immediately up to five times,
and only show the error to the user if that doesn’t resolve the problem.
Most of the time, this should be transparent to the user.