Introduce command result for maxlag error
This command record can be returned by Runner to indicate that the API
responded with an error that replication lag is currently too high. For
the time being, we just show the error and don’t offer any facilities
for trying again later, though I plan to add that in the future.