Introduce command result for blocked error
This command record can be returned by Runner if the API reports that
the user is blocked, or that the IP address we use was blocked
automatically because it was used by a blocked user.

Unfortunately, while the API returns a blockinfo object with detailed
information about the block, mwapi’s APIError class hides that from us,
so for now the blockinfo member is marked Optional in the type system
and always None in practice. Perhaps I’ll fix this in the future, either
by working around (or improving) mwapi or by making a separate API
request to get the blockinfo, but for now that’s not a priority.