Normalize domain name in batch table
There are only so many domains we work with, and most batches will focus
on the same few most common ones, so I think it makes sense to normalize
the domain by extracting it into a separate table with just an ID
reference in the batch table.

A maintenance SQL script is provided to migrate existing installations
to the new schema.

The Python code is extracted into a separate class with configurable
table and column names, because I intend to similarly normalize a few
more bits of information soon. It’s loosely inspired by MediaWiki’s
NameTableStore [1], but we don’t use it for reading yet, using a plain
SQL JOIN instead.

The “thinking face” emoji is used in one of the tests not just because
it’s funny, but because it’s outside the Basic Multilingual Plane [2].
(But also because it’s funny.)

[1]: https://doc.wikimedia.org/mediawiki-core/master/php/classMediaWiki_1_1Storage_1_1NameTableStore.html
[2]: https://twitter.com/LucasWerkmeistr/status/1029806980112756737