Cache string IDs in StringTableStore
Cache the IDs for the least recently used strings, so we don’t need to
ask the database for them all the time. A cache size of 1024 should be
enough to hold virtually all domains; perhaps I’ll make a larger cache
size configurable once we start to use this for normalizing other
strings, though I’m not sure if it’ll be necessary.