Split and partially normalize command_tpsv
The command_tpsv column of the command table, which used to hold the
full command of a command record, is split into the page and the
actions. The page is usually different for each command, and stored
directly in a column; the actions, on the other hand, will very often be
the same across an entire batch, and even between different batches, so
by normalizing those via a separate table, we can save a lot of space.