Add two timestamps to OpenBatch
An open batch now has two timestamps: when it was created, and when it
was last updated (i. e. one of its command records modified). It’s not
used anywhere yet, that will be done later.

This schema change cannot be migrated automatically, since the
information isn’t available in the database; for the Toolforge install,
I’ll manually set the timestamp based on the last edit in the batch and
server access log times, and for my local development install I’m fine
with just falling back to the Unix epoch. I’m not aware of any
third-party installs.

The database store updates the “last updated” timestamp in the database
each time a command record is modified, but the timestamps for OpenBatch
objects returned by previous get_batch() calls are not modified at the
moment. This should be okay for the time being, since the tool
frequently reloads batches anyways, but I’d like to fix it eventually
(though I’m not sure how).