Show batch timestamps on batch page
Show when a batch was created and last updated on the batch page. For
the text of the <time> element, we use the humanize library, which
returns nice strings like “1 hour ago” or “2 days ago”; unfortunately,
it doesn’t support timezone-aware datetimes, so we have to convert to
the local timezone and then remove the timezone first.