Fix test for DatabaseStore last_updated timestamp
After the removal of sub-second precision for these timestamps (commit
8ad626f1c9 and 07d8c6c83b), the assertion that last_updated > created is
no longer guaranteed to hold even when the test is correct. It seems to
be reliable on my local install, where MariaDB is abysmally slow for
reasons I can’t figure out [1], but is more prone to breaking on Travis
CI, where MariaDB seems to perform reasonably well. Let’s introduce a
one-second sleep before the update to force the timestamp to be higher
than the batch creation date.

[1]: https://twitter.com/LucasWerkmeistr/status/1107248901441159170