Fix errors on nochange edit response
If the wikitext we send to the edit API is the same as the current text,
it returns a 'nochange' response instead of a regular edit response,
which would cause a KeyError on the missing oldrevid. (The most likely
scenario for this is that the user accidentally included the same
command twice, so the second one represents no change from the first
one. This happened, I believe, in batch #21 [1].) Detect this case and
return a no-op instead.

In the scenario given above, updating the prepared_page with the
response from the first edit should also have prevented the error,
causing it to be detected before even calling the edit API, so I’m
adding a TODO to perhaps do this later. However, handing 'nochange'
responses from the edit API is still a good idea regardless, so let’s do
that first.

[1]: https://tools.wmflabs.org/quickcategories/batch/21/?offset=900&limit=50