Introduce CommandPending type
A CommandPending contains the same information as a CommandPlan, but is
a distinct status from it. Batches start out with a list of
CommandPlans, Runner only accepts CommandPendings, and the
BatchCommandRecords is responsible for atomically transitioning from one
to the other so that at most one execution unit will ever grab a pending
version of a plan even if multiple try to start running a batch in
parallel. The database version does this via a FOR UPDATE lock in SQL,
and the in-memory version doesn’t bother because all of InMemoryStore
isn’t concurrency-safe anyways.

This is mostly in preparation for running batches in the background, but
I think the race condition this fixes could in principle also have been
triggered by the user running the same batch in multiple browser tabs,
as long as the WSGI implementation was configured to use multiple
execution units (on Toolforge, uWSGI defaults to 4 worker processes).