Distinguish between open and closed batches
Two new batch classes are introduced: ClosedBatch, for a batch where all
commands are finished, and StoredBatch, a parent class of both OpenBatch
and StoredBatch. The stores are responsible for closing batches when the
last command is finished; the InMemoryStore checks this at batch
retrieval time to keep the BatchCommandRecordsList implementation
simple, whereas the DatabaseStore checks it when storing a command
finish, setting the status column to a new _BATCH_STATUS_CLOSED value if
there are no non-finished commands left in this batch. To make this
efficient, the previous command index on just the command batch is
replaced with one on both the command batch and status; the maintenance
script to update an existing database also sets the status of all
existing batches appropriately.

We don’t do anything with this information yet, but I think this
distinction will become important once we introduce background
processing of batches.