Add BatchStore.make_plan_pending_background() method
This method finds the batch that was not updated for the longest time
where an unstopped background run is available, and grabs its first
planned command, making it pending. It returns the associated batch, the
command itself, and a session with authentication data acquired from the
background run; parts of the session have to be supplied by the caller,
since they’re not stored in the database (and shouldn’t be: it’s a good
thing that the user OAuth secrets in the database are useless without
the consumer token that’s stored elsewhere).

This can be used to run batches in the background: the runner would
repeatedly call this method, run the returned command, and store it.
Since storing the finished command updates the batch’s timestamp, this
will result in a round-robin distribution over the batches, providing
some fairness (though without regard to which user submitted the batch,
or how many commands are remaining in it).

This only returns a single command so that stopping a batch has an
effect as soon as possible. However, if this turns out to severely
affect performance (we can no longer prefetch page revisions in
batches), I may change that and make the method return a slice of
commands. Perhaps it would make sense to ask users about this as well.