Run some DatabaseStore tests against InMemoryStore too
This introduces a `store` fixture which is parameterized to yield
different types of stores, and updates some tests to use that fixture
instead of instantiating a DatabaseStore directly.

Some tests are not updated because they also use the database directly;
some others are not updated because they currently fail for
InMemoryStore (it doesn’t update batch.last_updated properly yet).