Add background_runs accessor to batch classes
BatchBackgroundRuns is similar to BatchCommandRecords in that it doesn’t
directly store the information, it just gives you ways to access it,
which may load it from the database on-demand (though there’s no paging
this time – I figure it should rarely be necessary, since batches will
have way less background runs than command records).

There are methods to get only the latest background run or to get all of
them. Additionally, there’s a method to check if the batch is currently
running in the background; by default, this just checks if the latest
background run was stopped or not, but in the database implementation a
more efficient implementation is provided which the query server should
be able to answer only from an index (loading the full background run,
on the other hand, would require a table access).

I’m not adding tests right now because the store tests are in a bit of a
weird state at the moment; hopefully I’ll add them later.