Reorder joins in DatabaseStore.make_plan_pending_background()
As more and more background runs are started and stopped over the
lifetime of the tool, we expect the ratio of background runs matching
the `background_stopped_utc_timestamp` IS NULL condition to grow ever
smaller and smaller, so we definitely want MariaDB to start with the
background_stopped_batch index and apply that condition first. However,
when I tried out an EXPLAIN on the query locally, MariaDB would actually
start with the batch table (with no particular condition) and then join
the background table onto it with the background_batch_stopped index.
Rearranging the query so that we select from background and join batch
onto it seems to fix that, which hopefully improves query performance.

(Perhaps MariaDB would also have arrived at the more efficient plan on
its own given enough rows in the background table? I only have seven
background runs at the moment.)