Run background_runner in a loop
This is a very simple version of a continuous runner process: while
there are possible commands to run in the background, it keeps running
them as fast as possible, otherwise it sleeps for ten seconds before
asking for the next one. This should be an acceptable compromise between
not having the user wait for too long before commands start running, and
avoiding excessive load on the database server.

In the future, perhaps an improved messaging system can be used for the
app to inform the runner when there’s more work to do (avoiding the SQL
query every ten seconds). The Toolforge documentation [1] suggests using
Celery for long-running tasks triggered by a web frontend, with Redis as
a broker between worker and frontend – I don’t think Celery itself works
well for this tool, but Redis might be useful.

[1]: https://wikitech.wikimedia.org/wiki/Help:Toolforge