Run Travis CI on Python 3.5 as well
I want to upgrade to this Python version as soon as possible (Python 3.4
reached end-of-life), so let’s already ensure it works.

Since this means two jobs will run at the same time, we have to make
sure that test_runner.py doesn’t try to edit the page, otherwise the
jobs might step on each other’s feet. We get the Travis job number
(e. g. 82.2), get just the “sub-number” of this (e. g. 2), and use that
for the subpage to edit. Since the bot password used for these tests
doesn’t have the “create page” right, I had to create the subpages
manually; a note for future developers is added to .travis.yml.