Add assertions to _metadata_from_session
Ask the MediaWiki API to assert that the user is logged in, and locally
assert that the IDs and user name are nonempty.

Apparently this sometimes ends up being an anonymous request; I don’t
yet understand when that would happen, but these assertions seem like a
good idea anyways.