Check domain validity more thoroughly
When creating a new batch, first contact the API server and check that a
connection is possible, and that the server name reported by the API
server matches the domain name we contacted. This catches two kinds of
errors: domains like eu.wikimedia.org (instead of eu.wikipedia.org),
which don’t exist, and domains like commons.wikipedia.org (instead of
commons.wikimedia.org), which redirect to the real domain.