Don’t stop the batch when hitting a protected page
Previously, an error when failing to edit a protected page (even if the
user has the “editprotected” right, the OAuth consumer does not have the
“edit protected pages” grant) would not be caught, stopping the entire
batch. Now we detect it, properly record it, and continue with the rest
of the batch (since other pages are not affected by this error).