Forget about prepared page after edit
After a successful edit, we know that the prepared page must be out of
date. However, we don’t know the new wikitext: it’s not necessarily
exactly the wikitext we sent in our edit request, since MediaWiki may
have merged that with a non-conflicting edit that happened between our
preparation and run steps. Therefore, the safe thing to do is to forget
about this page, and fetch fresh information about it if it comes up
again (which is not expected to happen, but still).