Use LocalUser class for OpenBatch, ClosedBatch
This should make the user information a little easier to work with. The
domain is strictly speaking redundant with the user’s domain, but I
think it’s much clearer to keep it as a field directly on the batch too.

Later I’ll extend this into the database as well, but for the moment
this is just a change at the API level.