Show if background run was stopped by admin
If the last background run of a batch was stopped by another user than
the one who started it (currently, that means the stopping user must be
sysop on the target wiki), show that directly on the batch page. (The
full history page already showed this information.)

This is much nicer to implement now that we have the render_local_user()
function (though unfortunately it still introduces a fairly long line –
we can’t have any whitespace before the comma).