Extract local users into separate table
This joins the user name, local user ID and global user ID columns of
several tables into a single ID, deduplicating them along the way. As an
extra bonus, if a user is renamed, we update the record and all their
old batches will use the new user name as well (though this would still
need to be repeated for each wiki the user submitted a batch for).

This also allows for the future possibility of having background runs
stopped by users not local to the target wiki, e. g. global stewards.
Previously, this wasn’t possible because the domain to which the user
corresponded was implicitly the same as the batch domain – now it’s a
part of the localuser record and may differ from the batch domain.