Show “run in background” when current page is done
If the currently displayed page of commands is already done, the “run
commands” button is not shown, and previously that would hide the “run
in background” button as well. Now, that button is still shown as long
as the whole batch is not closed, because background runs are
independent of individual pages of commands.

The whole thing still needs to be one form so that the buttons appear
next to each other (the alternative would be two forms with forced
display: inline, I don’t want that), but the buttons are now more equal,
with the form as a whole no longer having an action that only
corresponds to one of the buttons. We also avoid evaluating the
can_run_commands() condition twice by saving the result in a local
variable (it’s not too expensive, but still).