Add reload hint for batches running in background
The batch page is static and doesn’t refresh itself while the batch runs
in the background, as the users might perhaps expect (e. g. from
QuickStatements). Add some text to indicate this, including a link that
users can click to reload.

The current_url() function is based on the one in Wikidata Lexeme Forms,
but adjusted so that it also preserves query arguments (offset, limit)
and non-external links (we don’t need absolute URLs).