Repeat certain failed commands later
Some command failures can’t be retried immediately but can be retried
later; the most important such type is “maxlag exceeded”, which can be
retried as soon as the wiki has recovered. We now append copies of such
commands to the end of the batch, where they can be picked up later.

The database backend also tracks the connection between original and
retried commands, though we don’t do anything with it for now. (I have
vague ideas for using this in the future, which will probably require at
least one if not two indexes on the table, but that can wait.)